Use prefix to provide default function name
This gives someone just getting started a simple starting point. It creates a unique name in their AWS environment for a single developer deployment.
If they want to create multiple lambdas in the same region they'll need to completely override this value, but this handles the simplest use case out
of the box.